requirements.txt
discord.py>=2.3
aiosqlite
apscheduler
requests
python-dotenv
feedparser
beautifulsoup4
